# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API is currently live at https://api.liteseed.xyz",
        "title": "Liteseed API",
        "contact": {
            "name": "API Support",
            "url": "https://liteseed.xyz/support",
            "email": "support@liteseed.xyz"
        },
        "version": "0.0.1"
    },
    "host": "https://api.liteseed.xyz",
    "paths": {
        "/": {
            "get": {
                "description": "Get the current status of the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Get status of the server",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.StatusResponse"
                        }
                    }
                }
            }
        },
        "/price/{bytes}": {
            "get": {
                "description": "Get the current price of data upload using the Liteseed Network.\nIt returns the price of upload in wei and the address to pay.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Get price of upload",
                "parameters": [
                    {
                        "maximum": 2147483647,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Size of Data",
                        "name": "bytes",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PriceGetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                }
            }
        },
        "/tx/": {
            "post": {
                "description": "Post your data in a specified ANS-104 data-item format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Post a data-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the data-item",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DataItemPostResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                }
            }
        },
        "/tx/{id}": {
            "get": {
                "description": "get all the fields of a posted data-item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Get a posted data-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the data-item",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                }
            }
        },
        "/tx/{id}/data": {
            "get": {
                "description": "get only the data of a posted data-item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Get data of a data-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the Data-Item",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "bytes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                }
            }
        },
        "/tx/{id}/status": {
            "get": {
                "description": "Get the current status of a posted data-item.\nResponse \"created\", \"queued\", \"sent\", \"confirmed\", \"failed\", \"invalid\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Get the status of a data-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the data-item",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                }
            }
        },
        "/tx/{id}/{payment_id}": {
            "put": {
                "description": "Once a payment is made send a transaction id for a data-item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Send a payment id for a data-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Data-Item id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment id",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.DataItemPutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.DataItemPostResponse": {
            "type": "object",
            "properties": {
                "dataCaches": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "deadlineHeight": {
                    "type": "integer"
                },
                "fastFinalityIndexes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "server.DataItemPutResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "string"
                }
            }
        },
        "server.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "integer"
                },
                "message": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "server.PriceGetResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "format": "string",
                    "example": "Cbj95zDZBBhmyht6iFlEf7xmSCSVZGw436V6HWmm9Ek"
                },
                "price": {
                    "type": "string",
                    "format": "string",
                    "example": "1000000000000"
                }
            }
        },
        "server.StatusResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "format": "string",
                    "example": "transit"
                },
                "version": {
                    "type": "string",
                    "format": "string",
                    "example": "v0.0.1"
                }
            }
        }
    }
}